+++ ts/src/coder.ts
@@ -1,4 +1,5 @@
 import camelCase from "camelcase";
+import * as base64 from "base64-js";
 import { snakeCase } from "snake-case";
 import { Layout } from "buffer-layout";
 import * as sha256 from "js-sha256";
@@ -213,6 +214,11 @@ class EventCoder {
    */
   private layouts: Map<string, Layout>;
 
+  /**
+   * Maps base64 encoded event discriminator to event name.
+   */
+  private discriminators: Map<string, string>;
+
   public constructor(idl: Idl) {
     if (idl.events === undefined) {
       this.layouts = new Map();
@@ -232,18 +238,29 @@ class EventCoder {
     });
     // @ts-ignore
     this.layouts = new Map(layouts);
+
+    this.discriminators = new Map<string, string>(
+      idl.events === undefined
+        ? []
+        : idl.events.map((e) => [
+            base64.fromByteArray(eventDiscriminator(e.name)),
+            e.name,
+          ])
+    );
   }
 
-  public encode<T = any>(eventName: string, account: T): Buffer {
-    const buffer = Buffer.alloc(1000); // TODO: use a tighter buffer.
-    const layout = this.layouts.get(eventName);
-    const len = layout.encode(account, buffer);
-    return buffer.slice(0, len);
-  }
+  public decode<T = any>(log: string): T | null {
+    const logArr = Buffer.from(base64.toByteArray(log));
+    const disc = base64.fromByteArray(logArr.slice(0, 8));
 
-  public decode<T = any>(eventName: string, ix: Buffer): T {
+    // Only deserialize if the discriminator implies a proper event.
+    const eventName = this.discriminators.get(disc);
+    if (eventName === undefined) {
+      return undefined;
+    }
+
     const layout = this.layouts.get(eventName);
-    return layout.decode(ix);
+    return layout.decode(logArr.slice(8));
   }
 }